--- NLayer.NAudioSupport/NLayer.NAudioSupport.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Authors>Mark Heath</Authors>
-    <Version>1.4.0</Version>
-    <PackageLicenseUrl></PackageLicenseUrl>
-    <PackageProjectUrl>https://github.com/naudio/NLayer</PackageProjectUrl>
-    <Description>NAudio support for NLayer</Description>
-    <PackageTags>NAudio NLayer MP3 audio</PackageTags>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NAudio.Core" Version="2.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\NLayer\NLayer.csproj" />
-  </ItemGroup>
-</Project>
