--- NLayer.sln
@@ -1,39 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.13
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestApp", "TestApp\TestApp.csproj", "{F133D458-0EC2-4612-8D14-89B788E4A0E9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NLayer.NAudioSupport", "NLayer.NAudioSupport\NLayer.NAudioSupport.csproj", "{33828999-1003-403C-BF78-59D9E02CE6A2}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AF5EF824-6615-4453-96A9-A89A69C59BF9}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLayer", "NLayer\NLayer.csproj", "{045D6A5F-BCA0-4144-AB8A-7AAC146420C9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F133D458-0EC2-4612-8D14-89B788E4A0E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F133D458-0EC2-4612-8D14-89B788E4A0E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F133D458-0EC2-4612-8D14-89B788E4A0E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F133D458-0EC2-4612-8D14-89B788E4A0E9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{33828999-1003-403C-BF78-59D9E02CE6A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{33828999-1003-403C-BF78-59D9E02CE6A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{33828999-1003-403C-BF78-59D9E02CE6A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{33828999-1003-403C-BF78-59D9E02CE6A2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{045D6A5F-BCA0-4144-AB8A-7AAC146420C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{045D6A5F-BCA0-4144-AB8A-7AAC146420C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{045D6A5F-BCA0-4144-AB8A-7AAC146420C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{045D6A5F-BCA0-4144-AB8A-7AAC146420C9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
