--- NLayer/NLayer.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard1.3;netstandard2.0</TargetFrameworks>
-    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Authors>Mark Heath, Andrew Ward</Authors>
-    <Version>1.16.0</Version>
-	<PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/naudio/NLayer</PackageProjectUrl>
-    <Description>Fully Managed MPEG 1 &amp; 2 Decoder for Layers 1, 2, &amp; 3</Description>
-    <PackageTags>audio csharp mp3</PackageTags>
-	<PackageReadmeFile>README.md</PackageReadmeFile>
-    <Company />
-  </PropertyGroup>
-
-	<ItemGroup>
-		<None Include="..\README.md" Pack="true" PackagePath="\" />
-	</ItemGroup>
-</Project>
