+++ XLayer/Decoder/RiffHeaderFrame.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace NLayer.Decoder
+namespace XLayer.Decoder
 {
     /// <summary>
     /// RIFF header reader
