+++ XLayer/Decoder/VBRInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace NLayer.Decoder
+﻿namespace XLayer.Decoder
 {
     class VBRInfo
     {
