+++ XLayer/Enums.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace NLayer
+namespace XLayer
 {
     public enum MpegVersion
     {
